fix mock data insertion to the db and separate it from db connection file
